Extract site preview with a border/shadow to a separate view `SitePreviewView`.
